refactor index.js
