+++ package.json
@@ -11,6 +11,7 @@
   "dependencies": {
     "@gunjanghate/git-genie": "^1.0.16",
     "@hookform/resolvers": "^3.10.0",
+    "@mistralai/mistralai": "^1.14.0",
     "@radix-ui/react-accordion": "1.2.2",
     "@radix-ui/react-alert-dialog": "1.1.4",
     "@radix-ui/react-aspect-ratio": "1.1.1",
@@ -49,6 +50,7 @@
     "date-fns": "4.1.0",
     "embla-carousel-react": "8.5.1",
     "geist": "latest",
+    "groq-sdk": "^0.37.0",
     "gsap": "^3.13.0",
     "input-otp": "1.4.1",
     "lenis": "^1.3.11",
